initial manifest dialect implementation
This commit mainly focus on defining the basic structure of the "manifest"
dialect, which is the simplest, and the most verbose configuration file.The
main purpose of this "manifest" dialect specification is to provide an
interchangable format that we can build more human-friendly dialects on top of
it, e.g., "benchmark", "pkgtest".

This commit includes the read and load of the TOML configuration file. The
execution of the workflow is not included here.