remove manifest order and PipelineOrder
Because pipeline order isn't generic enough to represent the task dependencies.
It turns out hard (sometimes impossible) to maintain the runtime `order`
configuration if we want to support general workflow oeprations such as
`filter`, `append!`. In contrast, a DAG represented by adjancency matrix
gives us pretty good flexibility.

This commit replaces the PipelineOrder and "order" entry with the newly
added TaskGraph object. The good (and bad) outcome of this is that users
are not impossible to specify the global orders.

Tests and examples are not changed in this commit.