CI: run all examples (#27)
The "examples/" directory is reorganized a bit with some fixes to ensure
all examples are runnable in new setup.

README.md and run.jl are added to improve the discoverbility.

Examples CI only runs on ubuntu to ensure all examples are correctly
written. Cross-platform and functionalities are checked by unit tests.